[#4] refactor: global 패키지로 이동
