--- src/main/java/com/beat/common/dto/SuccessResponse.java
@@ -1,17 +1,0 @@
-package com.beat.common.dto;
-
-import com.beat.exception.message.SuccessMessage;
-
-public record SuccessResponse<T>(
-        int status,
-        String message,
-        T data
-) {
-    public static <T> SuccessResponse of(final SuccessMessage successMessage, final T data) {
-        return new SuccessResponse(successMessage.getStatus(), successMessage.getMessage(), data);
-    }
-
-    public static SuccessResponse of(final SuccessMessage successMessage) {
-        return new SuccessResponse(successMessage.getStatus(), successMessage.getMessage(), null);
-    }
-}
