--- src/main/java/com/beat/exception/BadRequestException.java
@@ -1,9 +1,0 @@
-package com.beat.exception;
-
-import com.beat.exception.message.ErrorMessage;
-
-public class BadRequestException extends BeatException {
-    public BadRequestException(final ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
