--- src/main/java/com/beat/exception/ForbiddenException.java
@@ -1,9 +1,0 @@
-package com.beat.exception;
-
-import com.beat.exception.message.ErrorMessage;
-
-public class ForbiddenException extends BeatException {
-    public ForbiddenException(final ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
