--- src/main/java/com/beat/exception/NotFoundException.java
@@ -1,9 +1,0 @@
-package com.beat.exception;
-
-import com.beat.exception.message.ErrorMessage;
-
-public class NotFoundException extends BeatException {
-    public NotFoundException(final ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
