--- src/main/java/com/beat/exception/UnauthorizedException.java
@@ -1,9 +1,0 @@
-package com.beat.exception;
-
-import com.beat.exception.message.ErrorMessage;
-
-public class UnauthorizedException extends BeatException {
-    public UnauthorizedException(final ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
