+++ src/main/java/com/beat/global/common/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.beat.config;
+package com.beat.global.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
