+++ src/main/java/com/beat/global/common/exception/BeatException.java
@@ -1,0 +1,14 @@
+package com.beat.global.common.exception;
+
+import com.beat.global.common.exception.base.BaseErrorCode;
+import lombok.Getter;
+
+@Getter
+public class BeatException extends RuntimeException {
+    private final BaseErrorCode baseErrorCode;
+
+    public BeatException(BaseErrorCode baseErrorCode) {
+        super(baseErrorCode.getMessage());
+        this.baseErrorCode = baseErrorCode;
+    }
+}
