+++ src/main/java/com/beat/global/common/exception/ForbiddenException.java
@@ -1,0 +1,9 @@
+package com.beat.global.common.exception;
+
+import com.beat.global.common.exception.base.BaseErrorCode;
+
+public class ForbiddenException extends BeatException {
+    public ForbiddenException(final BaseErrorCode baseErrorCode) {
+        super(baseErrorCode);
+    }
+}
