+++ src/main/java/com/beat/global/common/exception/NotFoundException.java
@@ -1,0 +1,9 @@
+package com.beat.global.common.exception;
+
+import com.beat.global.common.exception.base.BaseErrorCode;
+
+public class NotFoundException extends BeatException {
+    public NotFoundException(final BaseErrorCode baseErrorCode) {
+        super(baseErrorCode);
+    }
+}
