+++ src/main/java/com/beat/global/common/exception/UnauthorizedException.java
@@ -1,0 +1,9 @@
+package com.beat.global.common.exception;
+
+import com.beat.global.common.exception.base.BaseErrorCode;
+
+public class UnauthorizedException extends BeatException {
+    public UnauthorizedException(final BaseErrorCode baseErrorCode) {
+        super(baseErrorCode);
+    }
+}
