+++ src/main/java/com/beat/global/common/handler/GlobalExceptionHandler.java
@@ -1,10 +1,10 @@
-package com.beat.common.handler;
+package com.beat.global.common.handler;
 
-import com.beat.common.dto.ErrorResponse;
-import com.beat.exception.BadRequestException;
-import com.beat.exception.ForbiddenException;
-import com.beat.exception.NotFoundException;
-import com.beat.exception.UnauthorizedException;
+import com.beat.global.common.dto.ErrorResponse;
+import com.beat.global.common.exception.BadRequestException;
+import com.beat.global.common.exception.ForbiddenException;
+import com.beat.global.common.exception.NotFoundException;
+import com.beat.global.common.exception.UnauthorizedException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -18,22 +18,22 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(BadRequestException.class)
     public ResponseEntity<ErrorResponse> handleBadRequestException(final BadRequestException e) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ErrorResponse.of(e.getErrorMessage()));
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ErrorResponse.of(e.getBaseErrorCode()));
     }
 
     @ExceptionHandler(UnauthorizedException.class)
     public ResponseEntity<ErrorResponse> handleUnauthorizedException(final UnauthorizedException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ErrorResponse.of(e.getErrorMessage()));
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ErrorResponse.of(e.getBaseErrorCode()));
     }
 
     @ExceptionHandler(ForbiddenException.class)
     public ResponseEntity<ErrorResponse> handleForbiddenException(final ForbiddenException e) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ErrorResponse.of(e.getErrorMessage()));
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ErrorResponse.of(e.getBaseErrorCode()));
     }
 
     @ExceptionHandler(NotFoundException.class)
     protected ResponseEntity<ErrorResponse> handleNotFoundException(final NotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ErrorResponse.of(e.getErrorMessage()));
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ErrorResponse.of(e.getBaseErrorCode()));
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
